Moved API logic to useSurveySubmits and used localStorage to track if survey answers were submitted
