+++ bigger-shape-web/src/components/Survey.tsx
@@ -2,6 +2,7 @@ import { useState, useEffect } from "react";
 import InputField from "./InputField";
 import ProgressBar from "./ProgressBar";
 import { useAuth } from "../AuthContext";
+import { useSurveySubmit } from "../useSurveySubmit";
 
 const API_BASE_URL = import.meta.env.VITE_API_BASE_URL;
 
@@ -10,6 +11,7 @@ function Survey() {
   const [questions, setQuestions] = useState(null);
   const [selectedAnswer, setSelectedAnswer] = useState([]);
   const [page, setPage] = useState(0);
+  const { handleSubmit } = useSurveySubmit();
 
   useEffect(() => {
     fetch(`${API_BASE_URL}/api/v1/public/questions`)
@@ -31,81 +33,6 @@ function Survey() {
   }
   function increPage() {
     setPage((curPage) => (curPage < questions.length ? curPage + 1 : curPage));
-  }
-
-  // Returns the body of the HTTPRequest for /api/v1/auth/users/history in a JSON format
-  const getHistoryRequestBody = () => {
-    const submissionAnswers: {
-      answerContent: string;
-      questionOrder: number;
-    }[] = [];
-    for (let i = 0; i < sessionStorage.length; i++) {
-      const value = sessionStorage.getItem(`question-${i}`);
-      if (!value) {
-        continue;
-      }
-      const answerData = JSON.parse(value);
-      submissionAnswers.push({
-        answerContent: answerData.answerValue,
-        questionOrder: i + 1,
-      });
-    }
-    return JSON.stringify({
-      questionnaire: {
-        dateTaken: new Date().toISOString(),
-        riskScore: 5,
-      },
-      answers: submissionAnswers,
-    });
-  };
-
-  // Returns the body of the HTTPRequest for /api/v1/auth/users/history in a JSON format
-  const getRecommendationsRequestBody = () => {
-    const submissionAnswers: {
-      questionId: string;
-      answer: number;
-    }[] = [];
-    for (let i = 0; i < sessionStorage.length; i++) {
-      const value = sessionStorage.getItem(`question-${i}`);
-      if (!value) {
-        continue;
-      }
-      const answerData = JSON.parse(value);
-      submissionAnswers.push({
-        questionId: answerData.questionId,
-        answer: answerData.answerIndex,
-      });
-    }
-    return JSON.stringify({
-      answers: submissionAnswers,
-    });
-  };
-
-  function handleSubmit() {
-    if (auth?.session?.access_token) {
-      fetch(`${API_BASE_URL}/api/v1/auth/users/history`, {
-        method: "POST",
-        headers: {
-          "Content-Type": "application/json",
-          Authorization: `Bearer ${auth.session.access_token}`,
-        },
-        body: getHistoryRequestBody(),
-      }).catch((error) => console.error("Error:", error));
-    } else {
-      // Error handling
-    }
-    fetch(`${API_BASE_URL}/api/v1/public/submit-answers`, {
-      method: "POST",
-      headers: {
-        "Content-Type": "application/json",
-      },
-      body: getRecommendationsRequestBody(),
-    })
-      .then((response) => response.json())
-      .then((body) => {
-        sessionStorage.setItem("recommendations", JSON.stringify(body));
-      })
-      .then(() => (window.location.href = "/dashboard"));
   }
 
   return (
